# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LeadingQuestionResponse": {
				"properties": {
					"question_id": {
						"type": "string"
					},
					"section_id": {
						"type": "string"
					},
					"question": {
						"type": "string"
					}
				},
				"required": [
					"question_id",
					"section_id",
					"question"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_LeadingQuestionResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/LeadingQuestionResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_LeadingQuestionResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/LeadingQuestionResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateLeadingQuestionRequest": {
				"properties": {
					"section_id": {
						"type": "string"
					},
					"question": {
						"type": "string"
					},
					"answer": {
						"type": "string"
					}
				},
				"required": [
					"section_id",
					"question",
					"answer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateLeadingQuestionRequest": {
				"properties": {
					"question_id": {
						"type": "string"
					},
					"section_id": {
						"type": "string"
					},
					"question": {
						"type": "string"
					},
					"answer": {
						"type": "string"
					}
				},
				"required": [
					"question_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExerciseOptionResponse": {
				"properties": {
					"option_id": {
						"type": "string"
					},
					"exercise_id": {
						"type": "string"
					},
					"option_text": {
						"type": "string"
					},
					"is_correct": {
						"type": "boolean"
					}
				},
				"required": [
					"option_id",
					"exercise_id",
					"option_text",
					"is_correct"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ExerciseOptionResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ExerciseOptionResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ExerciseOptionResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/ExerciseOptionResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateExerciseOptionRequest": {
				"properties": {
					"exercise_id": {
						"type": "string"
					},
					"option_text": {
						"type": "string"
					},
					"is_correct": {
						"type": "boolean"
					}
				},
				"required": [
					"exercise_id",
					"option_text",
					"is_correct"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateExerciseOptionRequest": {
				"properties": {
					"option_id": {
						"type": "string"
					},
					"exercise_id": {
						"type": "string"
					},
					"option_text": {
						"type": "string"
					},
					"is_correct": {
						"type": "boolean"
					}
				},
				"required": [
					"option_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExerciseResponse": {
				"properties": {
					"exercise_id": {
						"type": "string"
					},
					"section_id": {
						"type": "string"
					},
					"question": {
						"type": "string"
					},
					"type_status": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"answer": {
						"type": "string"
					}
				},
				"required": [
					"exercise_id",
					"question",
					"type_status",
					"score",
					"answer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ExerciseResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ExerciseResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ExerciseResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/ExerciseResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateExerciseRequest": {
				"properties": {
					"section_id": {
						"type": "string"
					},
					"question": {
						"type": "string"
					},
					"type_status": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"answer": {
						"type": "string"
					}
				},
				"required": [
					"question",
					"type_status",
					"answer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateExerciseRequest": {
				"properties": {
					"exercise_id": {
						"type": "string"
					},
					"section_id": {
						"type": "string"
					},
					"question": {
						"type": "string"
					},
					"type_status": {
						"type": "string"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"answer": {
						"type": "string"
					}
				},
				"required": [
					"exercise_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-pg-orm-backend-ts",
		"version": "1.0.0",
		"description": "TypeScript Express.js backend with PostgreSQL and Sequelize ORM",
		"contact": {}
	},
	"paths": {
...
		"/leading-questions/list": {
			"post": {
				"operationId": "ListLeadingQuestions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_LeadingQuestionResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"预设问题表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/leading-questions/getById": {
			"post": {
				"operationId": "GetLeadingQuestionById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_LeadingQuestionResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"预设问题表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"question_id": {
										"type": "string"
									}
								},
								"required": [
									"question_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/leading-questions/add": {
			"post": {
				"operationId": "AddLeadingQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"预设问题表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLeadingQuestionRequest"
							}
						}
					}
				}
			}
		},
		"/leading-questions/update": {
			"post": {
				"operationId": "UpdateLeadingQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"预设问题表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateLeadingQuestionRequest"
							}
						}
					}
				}
			}
		},
		"/leading-questions/delete": {
			"post": {
				"operationId": "DeleteLeadingQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"预设问题表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"question_id": {
										"type": "string"
									}
								},
								"required": [
									"question_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/exercise-options/list": {
			"post": {
				"operationId": "ListExerciseOptions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ExerciseOptionResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"习题选项表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/exercise-options/getById": {
			"post": {
				"operationId": "GetExerciseOptionById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ExerciseOptionResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"习题选项表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"option_id": {
										"type": "string"
									}
								},
								"required": [
									"option_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/exercise-options/add": {
			"post": {
				"operationId": "AddExerciseOption",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"习题选项表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateExerciseOptionRequest"
							}
						}
					}
				}
			}
		},
		"/exercise-options/update": {
			"post": {
				"operationId": "UpdateExerciseOption",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"习题选项表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateExerciseOptionRequest"
							}
						}
					}
				}
			}
		},
		"/exercise-options/delete": {
			"post": {
				"operationId": "DeleteExerciseOption",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"习题选项表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"option_id": {
										"type": "string"
									}
								},
								"required": [
									"option_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/exercises/list": {
			"post": {
				"operationId": "ListExercises",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ExerciseResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"习题表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/exercises/getById": {
			"post": {
				"operationId": "GetExerciseById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ExerciseResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"习题表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"exercise_id": {
										"type": "string"
									}
								},
								"required": [
									"exercise_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/exercises/add": {
			"post": {
				"operationId": "AddExercise",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"习题表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateExerciseRequest"
							}
						}
					}
				}
			}
		},
		"/exercises/update": {
			"post": {
				"operationId": "UpdateExercise",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"习题表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateExerciseRequest"
							}
						}
					}
				}
			}
		},
		"/exercises/delete": {
			"post": {
				"operationId": "DeleteExercise",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"习题表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"exercise_id": {
										"type": "string"
									}
								},
								"required": [
									"exercise_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}